CXX-1052 Pull find::convert_all_modifiers() into new class in private/
