+++ src/mongocxx/test/CMakeLists.txt
@@ -45,6 +45,7 @@ set(mongocxx_test_sources
     options/insert.cpp
     options/modify_collection.cpp
     options/pool.cpp
+    options/private/rewriter.cpp
     options/update.cpp
     pool.cpp
     private/scoped_bson_t.cpp
